[#12240] Change heatmap code structure
