+++ web-frontend/src/main/v3/packages/ui/src/components/Heatmap/index.tsx
@@ -1,2 +1,1 @@
-export * from './HeatmapFetcher';
-export * from './HeatmapChart';
+export * from './Heatmap';
